Put Moe in his place.
