+++ nucypher/characters/chaotic.py
@@ -1,0 +1,40 @@
+from hendrix.experience import hey_joe
+
+from nucypher.characters.base import Character
+from nucypher.network.nodes import FleetStateTracker
+
+
+class Moe(Character):
+    """
+    A monitor (lizard?)
+    """
+
+    class MonitoringTracker(FleetStateTracker):
+        def record_fleet_state(self, *args, **kwargs):
+            new_state_or_none = super().record_fleet_state(*args, **kwargs)
+            if new_state_or_none:
+                checksum, new_state = new_state_or_none
+                hey_joe.send({checksum: self.abridged_state_details(new_state)}, "states")
+            return new_state_or_none
+
+    tracker_class = MonitoringTracker
+    _SHORT_LEARNING_DELAY = .5
+    _LONG_LEARNING_DELAY = 30
+    LEARNING_TIMEOUT = 10
+    _ROUNDS_WITHOUT_NODES_AFTER_WHICH_TO_SLOW_DOWN = 25
+
+    def remember_node(self, *args, **kwargs):
+        new_node_or_none = super().remember_node(*args, **kwargs)
+        if new_node_or_none:
+            hey_joe.send(
+                {new_node_or_none.checksum_public_address: Moe.MonitoringTracker.abridged_node_details(new_node_or_none)},
+                "nodes")
+        return new_node_or_none
+
+    def learn_from_teacher_node(self, *args, **kwargs):
+        teacher = self.current_teacher_node(cycle=False)
+        new_nodes = super().learn_from_teacher_node(*args, **kwargs)
+        hey_joe.send({teacher.checksum_public_address: Moe.MonitoringTracker.abridged_node_details(teacher)}, "nodes")
+        new_teacher = self.current_teacher_node(cycle=False)
+        hey_joe.send({"current_teacher": new_teacher.checksum_public_address}, "teachers")
+        return new_nodes
