+++ nucypher/cli/moe.py
@@ -4,61 +4,14 @@ import os.path
 import click
 from constant_sorrow import constants
 from flask import Flask, render_template
-from twisted.logger import globalLogPublisher
-
 from hendrix.deploy.base import HendrixDeploy
 from hendrix.experience import hey_joe
-from nucypher.characters.base import Character
+
+from nucypher.characters.banners import MOE_BANNER
+from nucypher.characters.chaotic import Moe
 from nucypher.characters.lawful import Ursula
 from nucypher.cli.types import NETWORK_PORT
 from nucypher.network.middleware import RestMiddleware
-from nucypher.network.nodes import FleetStateTracker
-from nucypher.utilities.logging import SimpleObserver
-
-MOE_BANNER = r"""
- _______               
-|   |   |.-----..-----.
-|       ||  _  ||  -__|
-|__|_|__||_____||_____|
-
-the Monitor.
-"""
-
-
-class MonitoringTracker(FleetStateTracker):
-    def record_fleet_state(self, *args, **kwargs):
-        new_state_or_none = super(MonitoringTracker, self).record_fleet_state(*args, **kwargs)
-        if new_state_or_none:
-            checksum, new_state = new_state_or_none
-            hey_joe.send({checksum: self.abridged_state_details(new_state)}, "states")
-        return new_state_or_none
-
-
-class Moe(Character):
-    """
-    A monitor (lizard?)
-    """
-    tracker_class = MonitoringTracker
-    _SHORT_LEARNING_DELAY = .5
-    _LONG_LEARNING_DELAY = 30
-    LEARNING_TIMEOUT = 10
-    _ROUNDS_WITHOUT_NODES_AFTER_WHICH_TO_SLOW_DOWN = 25
-
-    def remember_node(self, *args, **kwargs):
-        new_node_or_none = super().remember_node(*args, **kwargs)
-        if new_node_or_none:
-            hey_joe.send(
-                {new_node_or_none.checksum_public_address: MonitoringTracker.abridged_node_details(new_node_or_none)},
-                "nodes")
-        return new_node_or_none
-
-    def learn_from_teacher_node(self, *args, **kwargs):
-        teacher = self.current_teacher_node(cycle=False)
-        new_nodes = super().learn_from_teacher_node(*args, **kwargs)
-        hey_joe.send({teacher.checksum_public_address: MonitoringTracker.abridged_node_details(teacher)}, "nodes")
-        new_teacher = self.current_teacher_node(cycle=False)
-        hey_joe.send({"current_teacher": new_teacher.checksum_public_address}, "teachers")
-        return new_nodes
 
 
 @click.command()
